requirement.txt
jupyter
jupyterthemes
numpy
gym
keras
tensorflow-gpu
pyglet
pyopengl
pyvirtualdisplay
